Replace StrExt with inherent str methods in libcore
