+++ src/test/compile-fail/single-primitive-inherent-impl.rs
@@ -15,7 +15,7 @@
 #![no_std]
 
 // OK
-#[lang = "str"]
+#[lang = "str_alloc"]
 impl str {}
 
 impl str {
